fix: 컴포넌트 분리&디자인 수정 #70
